+++ src/lib.rs
@@ -132,4 +132,4 @@ mod links;
 mod common;
 
 pub use auth::{Token, request_token, authorize_url, authenticate_url, access_token, verify_tokens};
-pub use common::{Response, ResponseIter, WebResponse};
+pub use common::{Response, ResponseIter, ResponseIterRef, ResponseIterMut, WebResponse};